feat: alterando a arquitetura dos usuários
