+++ backend/src/repositories/users/all-users-repository.ts
@@ -1,0 +1,16 @@
+import { PrismaClient } from "@prisma/client";
+
+const prisma = new PrismaClient();
+
+const allUsersRepository = async () => {
+  return await prisma.user.findMany({
+    select: {
+      id: true,
+      name: true,
+      email: true,
+      phone: true,
+    },
+  });
+};
+
+export default allUsersRepository;
